requirements.txt
Flask==2.3.3
requests==2.31.0
Werkzeug==2.3.7
openai>=1.32.0
python-dotenv==1.0.0
Flask-Session==0.5.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
